Added New Functionality
Now have the ability to write cover art to MP3 files and specify where to send MP3 files similarly to MP4 files. 

Now when prompted the user will enter where to send the video file and where to send the audio file. This has caused a change in the formatting of the DownloadTo.txt file.

the -l option will list destinations for MP4 and MP3s

Added a new error code and message